pyproject.toml
[tool.poetry]
name = "rose-gold"
version = "0.1.0"
description = ""
authors = ["Mwapsam <mwapsam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
wagtail = "6.0.2"
psycopg2-binary = "2.9.9"
django-environ = "0.11.2"
boto3 = "1.34.103"
anyascii = "0.3.2"
asgiref = "3.8.1"
beautifulsoup4 = "4.12.3"
botocore = "1.34.103"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
defusedxml = "0.7.1"
django = "5.0.4"
django-filter = "23.5"
django-modelcluster = "6.3"
django-permissionedforms = "0.1"
django-social-share = "2.3.0"
django-storages = "1.14.3"
django-taggit = "5.0.1"
django-treebeard = "4.7.1"
djangorestframework = "3.15.1"
draftjs-exporter = "5.0.0"
et-xmlfile = "1.1.0"
facebook-sdk = "3.1.0"
filetype = "1.2.0"
future = "1.0.0"
html5lib = "1.1"
idna = "3.7"
jmespath = "1.0.1"
l18n = "2021.3"
laces = "0.1.1"
oauthlib = "3.2.2"
openpyxl = "3.1.2"
pillow = "10.3.0"
pillow-heif = "0.16.0"
psycopg = "3.1.19"
psycopg2 = "2.9.9"
python-dateutil = "2.9.0.post0"
python-twitter = "3.5"
pytz = "2024.1"
requests = "2.31.0"
requests-oauthlib = "2.0.0"
s3transfer = "0.10.1"
six = "1.16.0"
soupsieve = "2.5"
sqlparse = "0.4.4"
telepath = "0.3.1"
typing-extensions = "4.11.0"
tzdata = "2024.1"
urllib3 = "2.2.1"
wagtail-social-share = "0.1.4"
webencodings = "0.5.1"
willow = "1.8.0"
gunicorn = "^22.0.0"
whitenoise = "^6.6.0"
wagtail-metadata = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
